set -e in publishcatalog to exit on failure
Problem:
When the deploy key is invalid and git clone fails, the rest commands still execute

Solution:
Add set -e. Jenkins use it by default in shell step but it is not propagate into the script context. Also remove git diff-index so as to do commit regardless of the differences